pyproject.toml
[tool.poetry]
name = "nightcity"
version = "0.0.0"
description = "A collection of scripts and tools for DevOps tasks"
authors = ["Chris Pressland <mail@cpressland.io>"]
readme = "README.md"

[tool.poetry.scripts]
nightcity = "nightcity:cli"

[tool.poetry.dependencies]
python = "^3.12"
typer = { extras = ["all"], version = "^0.9.0" }
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
loguru = "^0.7.2"
kr8s = "^0.13.0"
pydantic = { extras = ["email"], version = "^2.6.0" }
sqlalchemy = "^2.0.25"
sqlalchemy-orm = "^1.2.10"
azure-storage-blob = "^12.19.0"
tenacity = "^8.2.3"
azure-identity = "^1.15.0"
pyyaml = "^6.0.1"
python-box = "^7.1.1"
azure-keyvault = "^4.2.0"
msgraph-sdk = "^1.1.0"
psycopg2-binary = "^2.9.9"
sentry-sdk = "^1.40.0"
paramiko = "^3.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.13"


[tool.ruff]
select = ["ALL"]
ignore = [
    "ANN101",
    "COM812",
    "D203",
    "D213",
    "D407",
    "FA102",
    "G004",
    "PLR0913",
    "S608",
    "D104",
]
line-length = 120

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
